LayoutTests/fast/regex/constructor.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This test checks use of the regexp constructor.");

var re = /abc/;

shouldBeTrue("re === RegExp(re)");
shouldBeTrue("re !== new RegExp(re)");
shouldBeTrue("re !== RegExp(re,'i')");
shouldBeTrue("re !== new RegExp(re,'i')");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
