Safari-611.1.10/LayoutTests/fast/regex/constructor-expected.txt

This test checks use of the regexp constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS re === RegExp(re) is true
PASS re !== new RegExp(re) is true
PASS re !== RegExp(re,'i') is true
PASS re !== new RegExp(re,'i') is true
PASS successfullyParsed is true

TEST COMPLETE
